refactor(tests): move parser unit tests out of integration directory
Extract 4 parser-only tests (no I/O, no marks) into
tests/pipeline/test_phase2_cli_generator_flag.py so they run as fast
unit tests. Keep only the 2 integration-marked tests in
tests/integration/.
