+++ tests/integration/test_phase2_cli_generator_flag.py
@@ -1,4 +1,4 @@
-"""Tests for the --generator CLI flag on phase2_cli."""
+"""Integration tests for the --generator CLI flag on phase2_cli."""
 
 import json
 from pathlib import Path
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 import pytest
 
-from autonomous_discovery.phase2_cli import build_parser, main
+from autonomous_discovery.phase2_cli import main
 
 _PREMISES = """\
 ---
@@ -38,27 +38,7 @@ OfNat.ofNat : Prop
 
 
 class TestGeneratorFlag:
-    """Tests for the --generator argument on phase2_cli."""
-
-    def test_parser_accepts_template(self) -> None:
-        parser = build_parser()
-        args = parser.parse_args(["--generator", "template"])
-        assert args.generator == "template"
-
-    def test_parser_accepts_ollama(self) -> None:
-        parser = build_parser()
-        args = parser.parse_args(["--generator", "ollama"])
-        assert args.generator == "ollama"
-
-    def test_parser_default_is_template(self) -> None:
-        parser = build_parser()
-        args = parser.parse_args([])
-        assert args.generator == "template"
-
-    def test_parser_rejects_invalid_generator(self) -> None:
-        parser = build_parser()
-        with pytest.raises(SystemExit):
-            parser.parse_args(["--generator", "invalid"])
+    """Integration tests for the --generator argument on phase2_cli."""
 
     @pytest.mark.integration
     def test_cli_with_template_generator(self, tmp_path: Path) -> None:
